Move the render stuff into _deps, where it belongs, right?
